+++ packages/core/src/hooks/useContractCall.ts
@@ -1,8 +1,11 @@
 import { Interface } from '@ethersproject/abi'
 import { useMemo } from 'react'
+import { ChainId } from '../constants'
+import { QueryParams } from '../constants/type/QueryParams'
 import { Falsy } from '../model/types'
+import { RawCall } from '../providers'
 import { useChainCalls } from './useChainCalls'
-import { ChainCall } from '../providers/chainState/callsReducer'
+import { useChainId } from './useChainId'
 
 function warnOnInvalidContractCall(call: ContractCall | Falsy) {
   console.warn(
@@ -10,7 +13,7 @@ function warnOnInvalidContractCall(call: ContractCall | Falsy) {
   )
 }
 
-function encodeCallData(call: ContractCall | Falsy): ChainCall | Falsy {
+function encodeCallData(call: ContractCall | Falsy, chainId: ChainId): RawCall | Falsy {
   if (!call) {
     return undefined
   }
@@ -19,13 +22,17 @@ function encodeCallData(call: ContractCall | Falsy): ChainCall | Falsy {
     return undefined
   }
   try {
-    return { address: call.address, data: call.abi.encodeFunctionData(call.method, call.args) }
+    return { address: call.address, data: call.abi.encodeFunctionData(call.method, call.args), chainId }
   } catch {
     warnOnInvalidContractCall(call)
     return undefined
   }
 }
 
+/**
+ * @public
+ * @deprecated Use {@link useCall} instead.
+ */
 export interface ContractCall {
   abi: Interface
   address: string
@@ -33,12 +40,35 @@ export interface ContractCall {
   args: any[]
 }
 
-export function useContractCall(call: ContractCall | Falsy): any[] | undefined {
-  return useContractCalls([call])[0]
+/**
+ * @public
+ * @deprecated Use {@link useCalls} instead.
+ */
+export function useContractCall(call: ContractCall | Falsy, queryParams: QueryParams = {}): any[] | undefined {
+  return useContractCalls([call], queryParams)[0]
 }
 
-export function useContractCalls(calls: (ContractCall | Falsy)[]): (any[] | undefined)[] {
-  const results = useChainCalls(calls.map(encodeCallData))
+/**
+ * @public
+ * @deprecated Use {@link useCall} instead.
+ */
+export function useContractCalls(
+  calls: (ContractCall | Falsy)[],
+  queryParams: QueryParams = {}
+): (any[] | undefined)[] {
+  const chainId = useChainId({ queryParams })
+
+  const rawCalls = useMemo(
+    () => calls.map((call) => (chainId !== undefined ? encodeCallData(call, chainId) : undefined)),
+    [
+      JSON.stringify(
+        calls.map((call) => call && { address: call.address.toLowerCase(), method: call.method, args: call.args })
+      ),
+      chainId,
+    ]
+  )
+
+  const results = useChainCalls(rawCalls)
 
   return useMemo(
     () =>
@@ -50,6 +80,6 @@ export function useContractCalls(calls: (ContractCall | Falsy)[]): (any[] | unde
         }
         return call && result ? (call.abi.decodeFunctionResult(call.method, result) as any[]) : undefined
       }),
-    [results]
+    [JSON.stringify(results)]
   )
 }