+++ packages/core/src/hooks/useContractFunction.ts
@@ -1,12 +1,16 @@
-import { addressEqual, TransactionOptions } from '../../src'
+import { TransactionOptions } from '../../src'
 import { Contract } from '@ethersproject/contracts'
-import { Web3Provider } from '@ethersproject/providers'
+import { JsonRpcProvider } from '@ethersproject/providers'
 import { useCallback, useState } from 'react'
 import { useEthers } from './useEthers'
 import { usePromiseTransaction } from './usePromiseTransaction'
 import { LogDescription } from 'ethers/lib/utils'
+import { ContractFunctionNames, Params, TypedContract } from '../model/types'
 
-export function connectContractToSigner(contract: Contract, options?: TransactionOptions, library?: Web3Provider) {
+/**
+ * @internal Intended for internal use - use it on your own risk
+ */
+export function connectContractToSigner(contract: Contract, options?: TransactionOptions, library?: JsonRpcProvider) {
   if (contract.signer) {
     return contract
   }
@@ -22,19 +26,26 @@ export function connectContractToSigner(contract: Contract, options?: Transactio
   throw new TypeError('No signer available in contract, options or library')
 }
 
-export function useContractFunction(contract: Contract, functionName: string, options?: TransactionOptions) {
+/**
+ * @public
+ */
+export function useContractFunction<T extends TypedContract, FN extends ContractFunctionNames<T>>(
+  contract: T,
+  functionName: FN,
+  options?: TransactionOptions
+) {
   const { library, chainId } = useEthers()
-  const { promiseTransaction, state } = usePromiseTransaction(chainId, options)
+  const { promiseTransaction, state, resetState } = usePromiseTransaction(chainId, options)
   const [events, setEvents] = useState<LogDescription[] | undefined>(undefined)
 
   const send = useCallback(
-    async (...args: any[]) => {
+    async (...args: Params<T, FN>): Promise<void> => {
       const contractWithSigner = connectContractToSigner(contract, options, library)
       const receipt = await promiseTransaction(contractWithSigner[functionName](...args))
       if (receipt?.logs) {
         const events = receipt.logs.reduce((accumulatedLogs, log) => {
           try {
-            return addressEqual(log.address, contract.address)
+            return log.address.toLowerCase() === contract.address.toLowerCase()
               ? [...accumulatedLogs, contract.interface.parseLog(log)]
               : accumulatedLogs
           } catch (_err) {
@@ -47,5 +58,5 @@ export function useContractFunction(contract: Contract, functionName: string, op
     [contract, functionName, options, library]
   )
 
-  return { send, state, events }
+  return { send, state, events, resetState }
 }