+++ packages/core/src/hooks/useNotifications.ts
@@ -5,11 +5,17 @@ import { useInterval } from './useInterval'
 import { useConfig } from '../providers/config/context'
 
 function getExpiredNotifications(notifications: Notification[], expirationPeriod: number) {
+  if (expirationPeriod === 0) {
+    return []
+  }
   const timeFromCreation = (creationTime: number) => Date.now() - creationTime
 
   return notifications.filter((notification) => timeFromCreation(notification.submittedAt) >= expirationPeriod)
 }
 
+/**
+ * @public
+ */
 export function useNotifications() {
   const { chainId, account } = useEthers()
   const { addNotification, notifications, removeNotification } = useNotificationsContext()