+++ packages/docs/src/examples/GettingStarted.tsx
@@ -1,0 +1,38 @@
+import React from 'react'
+import ReactDOM from 'react-dom'
+
+import {
+  Mainnet,
+  DAppProvider,
+  useEtherBalance,
+  useEthers,
+  Config,
+} from '@usedapp/core'
+import { formatEther } from '@ethersproject/units'
+import { getDefaultProvider } from 'ethers'
+
+const config: Config = {
+  readOnlyChainId: Mainnet.chainId,
+  readOnlyUrls: {
+    [Mainnet.chainId]: getDefaultProvider('mainnet'),
+  },
+}
+
+ReactDOM.render(
+  <DAppProvider config={config}>
+    <App />
+  </DAppProvider>,
+  document.getElementById('root')
+)
+
+function App() {
+  const { activateBrowserWallet, account } = useEthers()
+  const etherBalance = useEtherBalance(account)
+  return (
+    <div>
+      {!account && <button onClick={() => activateBrowserWallet()}>Connect</button>}
+      {account && <p>Account: {account}</p>}
+      {etherBalance && <p>Balance: {formatEther(etherBalance)}</p>}
+    </div>
+  )
+}